Move some of the lock and briefcase operations to management client
