+++ clients/imodels-client-authoring/src/operations/IModelsApiUrlFormatter.ts
@@ -4,19 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 import { IModelsApiUrlFormatter as ManamegentIModelsApiUrlFormatter } from "@itwin/imodels-client-management";
 
-import { GetLockListUrlParams } from "./lock/LockOperationParams";
-
 export class IModelsApiUrlFormatter extends ManamegentIModelsApiUrlFormatter {
   public getBaselineUrl(params: { iModelId: string }): string {
     return `${this.baseUrl}/${params.iModelId}/baselinefile`;
-  }
-
-  public getLockListUrl(params: {
-    iModelId: string;
-    urlParams?: GetLockListUrlParams;
-  }): string {
-    return `${this.baseUrl}/${params.iModelId}/locks${this.formQueryString({
-      ...params.urlParams,
-    })}`;
   }
 }
