+++ clients/imodels-client-authoring/src/operations/OperationParamExports.ts
@@ -9,3 +9,12 @@ export * from "./changeset/ChangesetOperationParams";
 export * from "./changeset-extended-data/ChangesetExtendedDataOperationsParams";
 export * from "./changeset-group/ChangesetGroupOperationParams";
 export * from "./lock/LockOperationParams";
+
+/** @deprecated Import {@link GetLockListParams} directly from `@itwin/imodels-client-management` instead. */
+export type { GetLockListParams } from "@itwin/imodels-client-management";
+/** @deprecated Import {@link GetLockListUrlParams} directly from `@itwin/imodels-client-management` instead. */
+export type { GetLockListUrlParams } from "@itwin/imodels-client-management";
+/** @deprecated Import {@link LockLevelFilter} directly from `@itwin/imodels-client-management` instead. */
+export type { LockLevelFilter } from "@itwin/imodels-client-management";
+/** @deprecated Import {@link ReleaseBriefcaseParams} directly from `@itwin/imodels-client-management` instead. */
+export type { ReleaseBriefcaseParams } from "@itwin/imodels-client-management";
