+++ clients/imodels-client-authoring/src/operations/briefcase/BriefcaseOperations.ts
@@ -13,7 +13,6 @@ import { OperationOptions } from "../OperationOptions";
 import {
   AcquireBriefcaseParams,
   BriefcaseProperties,
-  ReleaseBriefcaseParams,
 } from "./BriefcaseOperationParams";
 
 export class BriefcaseOperations<
@@ -45,24 +44,6 @@ export class BriefcaseOperations<
       params.headers
     );
     return result;
-  }
-
-  /**
-   * Releases the specified Briefcase. Wraps the
-   * {@link https://developer.bentley.com/apis/imodels-v2/operations/release-imodel-briefcase/ Release iModel Briefcase}
-   * operation from iModels API.
-   * @param {ReleaseBriefcaseParams} params parameters for this operation. See {@link ReleaseBriefcaseParams}.
-   * @returns {Promise<void>} a promise that resolves after operation completes.
-   */
-  public async release(params: ReleaseBriefcaseParams): Promise<void> {
-    await this.sendDeleteRequest({
-      authorization: params.authorization,
-      url: this._options.urlFormatter.getSingleBriefcaseUrl({
-        iModelId: params.iModelId,
-        briefcaseId: params.briefcaseId,
-      }),
-      headers: params.headers,
-    });
   }
 
   private getAcquireBriefcaseRequestBody(
