+++ clients/imodels-client-authoring/src/operations/lock/LockOperationParams.ts
@@ -2,29 +2,9 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import {
-  CollectionRequestParams,
-  IModelScopedOperationParams,
-} from "@itwin/imodels-client-management";
+import { IModelScopedOperationParams } from "@itwin/imodels-client-management";
 
-import { LockLevel, LockedObjects } from "../../base/types";
-
-/** Valid values for Lock filtering by their level. */
-export type LockLevelFilter = LockLevel.Shared | LockLevel.Exclusive;
-
-/** Url parameters supported in Lock list query. */
-export interface GetLockListUrlParams extends CollectionRequestParams {
-  /** Filters Locks by a specific Briefcase. */
-  briefcaseId?: number;
-  /** Filters Locks by their level. See {@link LockLevelFilter}. */
-  lockLevel?: LockLevelFilter;
-}
-
-/** Parameters for get Lock list operation. */
-export interface GetLockListParams extends IModelScopedOperationParams {
-  /** Parameters that will be appended to the entity list request url that will narrow down the results. */
-  urlParams?: GetLockListUrlParams;
-}
+import { LockedObjects } from "../../base/types";
 
 /** Parameters for update Lock operation. */
 export interface UpdateLockParams extends IModelScopedOperationParams {
