+++ clients/imodels-client-authoring/src/operations/lock/LockOperations.ts
@@ -3,42 +3,18 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import {
-  EntityListIterator,
-  EntityListIteratorImpl,
-  OperationsBase,
+  Lock,
+  LockResponse,
+  LockOperations as ManagementLockOperations,
 } from "@itwin/imodels-client-management";
 
-import { LockResponse, LocksResponse } from "../../base/internal";
-import { Lock } from "../../base/types";
 import { OperationOptions } from "../OperationOptions";
 
-import { GetLockListParams, UpdateLockParams } from "./LockOperationParams";
+import { UpdateLockParams } from "./LockOperationParams";
 
 export class LockOperations<
   TOptions extends OperationOptions
-> extends OperationsBase<TOptions> {
-  /**
-   * Gets Locks for a specific iModel. This method returns Locks in their full representation. The returned iterator
-   * internally queries entities in pages. Wraps the
-   * {@link https://developer.bentley.com/apis/imodels-v2/operations/get-imodel-locks/ Get iModel Locks} operation from
-   * iModels API.
-   * @param {GetLockListParams} params parameters for this operation. See {@link GetLockListParams}.
-   * @returns {EntityListIterator<Lock>} iterator for Lock list. See {@link EntityListIterator}, {@link Lock}.
-   */
-  public getList(params: GetLockListParams): EntityListIterator<Lock> {
-    return new EntityListIteratorImpl(async () =>
-      this.getEntityCollectionPage<Lock, LocksResponse>({
-        authorization: params.authorization,
-        url: this._options.urlFormatter.getLockListUrl({
-          iModelId: params.iModelId,
-          urlParams: params.urlParams,
-        }),
-        entityCollectionAccessor: (response) => response.body.locks,
-        headers: params.headers,
-      })
-    );
-  }
-
+> extends ManagementLockOperations<TOptions> {
   /**
    * Updates Lock for a specific Briefcase. This operation is used to acquire new locks and change the lock level for
    * already existing ones. Wraps the {@link https://developer.bentley.com/apis/imodels-v2/operations/update-imodel-locks/
