+++ clients/imodels-client-management/src/operations/OperationParamExports.ts
@@ -7,6 +7,7 @@ export * from "./briefcase/BriefcaseOperationParams";
 export * from "./changeset/ChangesetOperationParams";
 export * from "./changeset-extended-data/ChangesetExtendedDataOperationParams";
 export * from "./changeset-group/ChangesetGroupOperationParams";
+export * from "./lock/LockOperationParams";
 export * from "./named-version/NamedVersionOperationParams";
 export * from "./checkpoint/CheckpointOperationParams";
 export * from "./thumbnail/ThumbnailOperationParams";
