+++ clients/imodels-client-management/src/operations/lock/index.ts
@@ -2,8 +2,5 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { BaselineFile } from "../types";
-
-export interface BaselineFileResponse {
-  baselineFile: BaselineFile;
-}
+export * from "./LockOperationParams";
+export * from "./LockOperations";
