+++ common/changes/@itwin/imodels-client-management/move-lock-and-briefcase-release-to-management_2026-04-21.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/imodels-client-management",
+      "comment": "Add LockOperations (getList) and BriefcaseOperations.release, moved from @itwin/imodels-client-authoring",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@itwin/imodels-client-management"
+}
