+++ tests/imodels-clients-tests/src/integration/management/LockOperations.test.ts
@@ -1,0 +1,142 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import { expect } from "chai";
+
+import {
+  AuthorizationCallback,
+  GetLockListParams,
+  IModelsClient,
+  IModelsClientOptions,
+  Lock,
+  LockLevel,
+  LockLevelFilter,
+  toArray,
+} from "@itwin/imodels-client-management";
+import {
+  ReusableIModelMetadata,
+  ReusableTestIModelProvider,
+  TestAuthorizationProvider,
+  TestUtilTypes,
+  assertCollection,
+  assertLock,
+} from "@itwin/imodels-client-test-utils";
+
+import { getTestDIContainer } from "../common";
+
+describe("[Management] LockOperations", () => {
+  let iModelsClient: IModelsClient;
+  let authorization: AuthorizationCallback;
+
+  let testIModelForRead: ReusableIModelMetadata;
+
+  before(async () => {
+    const container = getTestDIContainer();
+
+    const iModelsClientOptions = container.get<IModelsClientOptions>(
+      TestUtilTypes.IModelsClientOptions
+    );
+    iModelsClient = new IModelsClient(iModelsClientOptions);
+
+    const authorizationProvider = container.get(TestAuthorizationProvider);
+    authorization = authorizationProvider.getAdmin1Authorization();
+
+    const reusableTestIModelProvider = container.get(
+      ReusableTestIModelProvider
+    );
+    testIModelForRead = await reusableTestIModelProvider.getOrCreate();
+  });
+
+  it("should return all items when querying collection", async () => {
+    // Arrange
+    const getLockListParams: GetLockListParams = {
+      authorization,
+      iModelId: testIModelForRead.id,
+    };
+
+    // Act
+    const locks = iModelsClient.locks.getList(getLockListParams);
+
+    // Assert
+    await assertCollection({
+      asyncIterable: locks,
+      isEntityCountCorrect: (count) => count === 1,
+    });
+  });
+
+  it("should return correct values when querying collection with briefcaseId filter", async () => {
+    // Arrange
+    const getLockListParams: GetLockListParams = {
+      authorization,
+      iModelId: testIModelForRead.id,
+      urlParams: {
+        briefcaseId: testIModelForRead.briefcases[0].id,
+      },
+    };
+
+    // Act
+    const locks = iModelsClient.locks.getList(getLockListParams);
+
+    // Assert
+    const lockArray = await toArray(locks);
+    const lock = lockArray[0];
+    assertLock({
+      actualLock: lock,
+      expectedLock: testIModelForRead.lock,
+    });
+  });
+
+  it("should return empty collection when querying locks for non-existent briefcase", async () => {
+    // Arrange
+    const getLockListParams: GetLockListParams = {
+      authorization,
+      iModelId: testIModelForRead.id,
+      urlParams: {
+        briefcaseId: 500,
+      },
+    };
+
+    // Act
+    const locks = iModelsClient.locks.getList(getLockListParams);
+
+    // Assert
+    const lockArray = await toArray(locks);
+    expect(lockArray.length).to.equal(0);
+  });
+
+  ([LockLevel.Shared, LockLevel.Exclusive] as LockLevelFilter[]).forEach(
+    (lockLevel) => {
+      it(`should return correct values when querying collection with lockLevel filter (${lockLevel})`, async () => {
+        // Arrange
+        const expectedLock: Lock = {
+          briefcaseId: testIModelForRead.lock.briefcaseId,
+          lockedObjects: testIModelForRead.lock.lockedObjects.filter(
+            (lockedObject) => lockedObject.lockLevel === lockLevel
+          ),
+        };
+        expect(expectedLock.lockedObjects.length).to.be.greaterThan(0);
+
+        const getLockListParams: GetLockListParams = {
+          authorization,
+          iModelId: testIModelForRead.id,
+          urlParams: {
+            lockLevel,
+          },
+        };
+
+        // Act
+        const locks = iModelsClient.locks.getList(getLockListParams);
+
+        // Assert
+        const lockArray = await toArray(locks);
+        expect(lockArray.length).to.be.equal(1);
+        const actualLock = lockArray[0];
+        assertLock({
+          actualLock,
+          expectedLock,
+        });
+      });
+    }
+  );
+});
